folder organization and use of favorite context in MoviesFavorites
